+++ LMeter/ACT/DataStructures/ActEvent.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using Dalamud.Plugin.Services;
 using LMeter.Helpers;
 using Newtonsoft.Json;
 
@@ -20,7 +21,6 @@ namespace LMeter.Act.DataStructures
         private bool _active;
 
         public DateTime Timestamp { get; set; } = DateTime.UtcNow;
-        public string Data { get; set; } = string.Empty;
 
         [JsonProperty("type")]
         public string EventType { get; set; } = string.Empty;
@@ -34,9 +34,9 @@ namespace LMeter.Act.DataStructures
         [JsonProperty("Combatant")]
         public Dictionary<string, Combatant>? Combatants { get; set; }
         
-        public string GetFormattedString(string format, string numberFormat)
+        public string GetFormattedString(string format, string numberFormat, bool emptyIfZero)
         {
-            return TextTagFormatter.TextTagRegex.Replace(format, new TextTagFormatter(this, numberFormat, _textTagMembers).Evaluate);
+            return TextTagFormatter.TextTagRegex.Replace(format, new TextTagFormatter(this, numberFormat, emptyIfZero, _textTagMembers).Evaluate);
         }
 
         public bool IsEncounterActive()
@@ -74,6 +74,38 @@ namespace LMeter.Act.DataStructures
             }
 
             return true;
+        }
+
+        public void InjectFFLogsData(FFLogsMeter? meter)
+        {
+            string playerName = CharacterState.CharacterName;
+            if (meter?.Actors is null ||
+                meter?.Encounter is null ||
+                this.Combatants is null ||
+                this.Encounter is null ||
+                string.IsNullOrEmpty(playerName))
+            {
+                return;
+            }
+
+            TimeSpan duration = TimeSpan.FromMilliseconds(meter.EncounterEnd - meter.EncounterStart - meter.Downtime);
+            foreach (Combatant combatant in this.Combatants.Values)
+            {
+                string name = combatant.OriginalName;
+                if (name.Equals("YOU"))
+                {
+                    name = playerName;
+                }
+
+                foreach (FFLogsActor actor in meter.Actors.Values)
+                {
+                    if (actor.Name.Equals(name))
+                    {
+                        combatant.FFLogsActor = actor;
+                        combatant.FFLogsDuration = duration;
+                    }
+                }
+            }
         }
 
         public static ActEvent GetTestData()