+++ LMeter/ACT/DataStructures/IActData.cs
@@ -9,6 +9,6 @@ namespace LMeter.Act.DataStructures
         static abstract string[] TextTags { get; }
         static abstract T GetTestData();
         
-        string GetFormattedString(string format, string numberFormat);
+        string GetFormattedString(string format, string numberFormat, bool emptyIfZero);
     }
 }